+++ Assets/Models.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: c2543b27dc0046b41a75dd47232c24bf
+guid: 3ab61ad659ddba14aa4737785bfc74f0
 folderAsset: yes
-timeCreated: 1508673652
-licenseType: Free
+timeCreated: 1508814054
+licenseType: Pro
 DefaultImporter:
   userData: 
   assetBundleName:
