+++ Assets/Prefabs.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d3d21bcc75a0c1843b4f7002d244dba5
+guid: c4613ccb74b11794f8dd6764782c6bc1
 folderAsset: yes
-timeCreated: 1508673619
-licenseType: Free
+timeCreated: 1508813922
+licenseType: Pro
 DefaultImporter:
   userData: 
   assetBundleName:
