+++ Assets/Scripts.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 525273361013636458c81cff7ec06e47
+guid: 4cb38241cd3280244bb3c23005c00646
 folderAsset: yes
-timeCreated: 1508673606
-licenseType: Free
+timeCreated: 1508813917
+licenseType: Pro
 DefaultImporter:
   userData: 
   assetBundleName:
